Rework GP fake server to have a persistent configuration
Until now, all of our Flask-based servers for testing protocol
authentication flows have "cheated" a bit.

They're configurable (e.g.  should the authentication require a 2FA token,
or should it not?), but this configuration has not actually been persisted
by the server.

Instead, the server sends the client a session cookie which has the desired
*server* configuration glommed into it, and relies on the client sending
this cookie back on every request (which OpenConnect obliges, for
authentication requests) so that the server can "remember" what it's
configured to do.

This is very confusing to read and understand, and it's also untenable to
continue using it for fake servers that simulate external authentication
(SAML/SSO).  That's because external authentication uses a separate
browser/handler for the authentication flow, so it won't share the session
cookies.

This modifies the fake GP server to store a persistent in-memory
configuration which can easily be set and inspected with cURL:

    $ ./fake-gp-server localhost 8080 certs/server-{cert,key}.pem 2>&1 >/dev/null &
    $ curl -sk https://localhost:8080/CONFIGURE -d gw_2fa=1
    $ curl -sk https://localhost:8080/CONFIGURE
    Current configuration of fake GP server configuration:
    TestConfiguration(gateways=('Default gateway',), portal_2fa=False, gw_2fa=True, portal_cookie=None, portal_saml=None, gateway_saml=None)
    $ openconnect --protocol=gp localhost:8080
    Please login to this fake GP VPN portal
    Username: fakeusername
    Password: *******
    ...
    2FA challenge from gateway
    Challenge:

Contrast this with the old approach:

    $ ./fake-gp-server localhost 8080 certs/server-{cert,kem}.pem 2>&1 >/dev/null &
    $ openconnect --protocol=gp localhost:8080/global-protect/testconfig.esp?gw_2fa=1

The gp-auth-and-config tests are modified accordingly to set the persistent
configuration of the fake server.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>