Don't install list-system-keys
It breaks the Linux COPR builds since the native RPM doesn't list it as
one of the files that are expected to be installed. And in fact we don't
*want* it installed for Linux as GnuTLS system keys aren't supported on
Linux anyway. So just leave it uninstalled for now. It's in the Windows
installer.

We *will* want to make it available for the GUI installer to use, but
I'm still not sure how to handle that. Since the GUI changes less often
perhaps I should actually pull in the openconnect-gui MinGW package as
a dependency of the openconnect build. Which is backwards really, but
it would allow us to spit out an installer including OpenConnect-GUI
for every libopenconnect build. Not today though...

Signed-off-by: David Woodhouse <dwmw2@infradead.org>