Clarify purpose/scope of --authgroup option
We frequently get questions from users who are unsure of how to
automatically enter an authentication dropdown selection using the command
line client. A recent example:
https://lists.infradead.org/pipermail/openconnect-devel/2022-May/005125.html

The `--authgroup=GROUP` option is specifically designed for this purpose: it can enter
a value into "the right" dropdown/list field on multiple protocols:

- Cisco AnyConnect/ocserv: "authgroup" selection form field
- Juniper: "realm" OR "frmSelectRoles" selection form field
- Pulse: "realm" selection form field
- Fortinet: "realm" selection form field
- F5: "domain" selection form field
- GlobalProtect: "gateway" selection form field (found on the "portal" interface;
  this one actually controls the choice of gateway server)

The functionality of the `--authgroup` option is not as obvious as
it could/should be because the name "authgroup" is Cisco-specific.

This patch improves the `--help` output and openconnect(8) man page to
show that it works with other protocols as well, and mention the names
of the relevant fields for those protocols.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>