Treat empty redirect_url as a no-op
If an HTML form has `action=""` or `action="#"`, this should be treated as a
no-op rather than as a redirect to the parent directory.

Per https://stackoverflow.com/a/37847231/20789, many HTML forms including
these empty(-ish) values because omitting the 'action' property was not
allowed prior to HTML5.

This should fix the issue seen with
F5 in https://gitlab.com/openconnect/openconnect/-/issues/464, and we can
also proactively make the same change for other protocols.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>