Do not add 'single-sign-on' to the capabilities list for AnyConnect auth requests
In 024336a8ddeb1754ae5e8fb18770e90c206070b1, we added 'single-sign-on-v2' to
the list of auth-methods capabilities in AnyConnect auth requests, and also
included 'single-sign-on' (no '-v2') because we had seen it included in a
MITM capture from a Cisco client.  See discussion at
https://gitlab.com/openconnect/openconnect/-/merge_requests/126#note_853084596.

However, OpenConnect does not actually know how to handle the
'single-sign-on' case, and include it appears to confuse Cisco servers and
cause them not to return the expected XML structure in their subsequent
responses.  See
https://gitlab.gnome.org/GNOME/NetworkManager-openconnect/-/issues/72 and
further discussion on
https://gitlab.com/openconnect/openconnect/-/merge_requests/394.

Removing 'single-sign-on' from the list of auth-methods capabilities should
resolve this issue.

Signed-off-by: Rahul Rameshbabu <sergeantsagara@protonmail.com>
Co-Authored-By: Eric Work <work.eric@gmail.com>
Signed-off-by: Daniel Lenski <dlenski@gmail.com>