Clarify certificate verification in Cisco CSD/trojan scripts
Environment variable CSD_SHA256 is always set by OpenConnect v8.00,
but was never set in earlier versions.

We expect cURL to verify server certificates against the public key in
CSD_SHA256. This requires cURL options '-k --pinnedpubkey'. Unfortunately,
'--pinnedpubkey; is not available until cURL v7.39.

For earlier versions of OpenConnect and/or cURL, we must either:

1. Skip certificate validate entirely with 'curl -k' (also known
   as 'curl --insecure').
2. Fallback to cURL's default certificate validation. This can result
   in failed verification in cases where the correct certificate
   public key has already been made known to OpenConnect, such as with
   'openconnect --servercert=...').

We should default to (2) since it is more secure, but allow the user to
choose (1) by explicitly setting the variable MISSING_OPTION_PINNEDPUBKEY,
which was previously known simply as INSECURE, in the source of the script.

Signed-off-by: Dimitri Papadopoulos <3350651-DimitriPapadopoulos@users.noreply.gitlab.com>
Signed-off-by: Daniel Lenski <dlenski@gmail.com>