Don't treat forms containing only hidden fields as non-empty
Filling an auth form that contains only hidden fields with fixed values
shouldn't cause that form to be treated as non-empty.

However, our 'juniper-sso-auth' tests include a case where there are 2
consecutive forms containing only hidden fields, and will fail with no other
change.  Instead, we relax the limit of consecutive empty forms to 3, before
deciding that we are stuck in an infinite loop.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>