Fortinet: fix bug causing segfault when SVPNCOOKIE is set repeatedly
As shown in https://gitlab.com/openconnect/openconnect/-/issues/514#note_1161271009 and
in the preceding commit c74b131c8b1fd34040a6c909fe5e4a2163f1330f ("Simulate
condition leading to segfault in fake-fortinet-server.py"),
some Fortinet servers re-set the `SVPNCOOKIE` repeatedly during the `fortinet_configure()`
flow.

Our existing code assumed incorrectly that the `SVPNCOOKIE` would be set
only once, while completing authentication, leading to a crash when building
the DTLS connection request (because the improperly saved reference to the
*original* cookie value had been `free()`ed).

The best solution is simply *not* to save a reference to the `SVPNCOOKIE`
value, but simply to defer looking up its value until we actually use it to
build the DTLS connection request.  This now aligns with what we do when
building the Fortinet *TLS* connection request, and with other protocols.

[ Manuel de Prada found and fixed the bug; DRL restructured a bit. ]

Signed-off-by: Daniel Lenski <dlenski@gmail.com>