Simulate condition leading to segfault in fake-fortinet-server.py
As shown in https://gitlab.com/openconnect/openconnect/-/issues/514#note_1161271009,
some Fortinet servers send the `Set-Cookie: SVPNCOOKIE` header not only in
response to successful authentication (`POST /remote/logincheck`, which is
when OpenConnect expects to receive it), but also in response to the
subsequent configuration request (`GET /remote/fortisslvpn_xml`):

    POST https://REDACTED_URL/remote/logincheck
    …
    Got HTTP response: HTTP/1.1 200 OK
    Set-Cookie: SVPNCOOKIE=<elided>; path=/; secure; httponly
    …
    GET https://REDACTED_URL/remote/fortisslvpn_xml
    …
    Got HTTP response: HTTP/1.1 200 OK
    Set-Cookie: SVPNCOOKIE=<elided>; path=/; secure; httponly
    …
    < … <fos platform='FG3K0B' major='5' minor='02' patch='7' build='0718' branch='718' /> …

When this happens, OpenConnect will call `free()` on the first copy of the
cookie's value, replacing it with a new value in `vpninfo->cookies`.  This
causes a segfault when it later tries to use a saved reference to the
*original* cookie value to establish the tunnel connection.

We should replicate that (mis)behaviour in our fake-fortinet-server.py, and
thus detect it in our tests.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>