Mention newer/non-PPP-based wire protocol in the Fortinet docs
This apparently-newer wire protocol has been discussed in
https://gitlab.com/openconnect/openconnect/-/issues/297#note_1173189545.

Thus far, we do not know of any advantages relative to the existing
PPP-based wire protocol, which remains available on all known Fortinet
servers.  If we discover some advantage of the new wire protocol — perhaps
less-broken handling of reconnection?  — then it should be relatively easy
to add support.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>