Parse GlobalProtect JavaScript challenge 'respMsg' as JSON string
We already have json-parser as a dependency (in order to support the Array
protocol), so we may as well use it to parse JavaScript strings for
GlobalProtect challenge messages. Yay?

JavaScript/JSON strings can be mangled in various ways (like `\u` followed
by anything other than 4 hex digits), so we still need the
naïve/non-JSON-aware handling as a fallback.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>