Replace deprecated libtasn1 macros
New macros were introduced in libtasn1 3.0 on 2012-10-28:
* `ASN1_TYPE` → `asn1_node`
* `ASN1_DATA_NODE` → `asn1_data_node_st`
* `ASN1_TYPE_EMPTY` → `NULL`

We would have to go back to RHEL 6 or Fedora 16 in order to find a version
of libtasn1 which is so old that it only supports the old names.

Because OpenConnect no longer tests against CentOS 6 in CI, and no longer
supports it, we can get rid of the deprecated types.

Signed-off-by: Dimitri Papadopoulos <3350651-DimitriPapadopoulos@users.noreply.gitlab.com>
Signed-off-by: Daniel Lenski <dlenski@gmail.com>