Give more details about unexpected Pulse configuration packets
Until now, we have logged the same catch-all message in the case of any
unexpected value in the header of the Pulse configuration packet:
"Unexpected IF-T/TLS packet when expecting configuration."

This is confusing to users and makes it hard to narrow down the actual
problem (see https://gitlab.com/openconnect/openconnect/-/issues/538#note_1220538728
and https://gitlab.com/openconnect/openconnect/-/issues/459#note_1220539246).

The configuration packet format seems to be an area of ongoing change in
recent Pulse server releases (see #58, #98, #379 → !331 and #472 → !401).
We should log more precise messages to aid in debugging and in adding
support for newer server versions.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>