cstp: Check if cookies is NULL in sso_detect_done
Before change if SSO implementers pass a NULL cookies array in
oc_webview_result, cstp_sso_detect_done would blindly dereference the NULL
pointer. This leads to panics that the implementer has to then trace inside of
openconnect. This NULL pointer check is similar to the one in
gpst_sso_detect_done for the headers array.

Signed-off-by: Rahul Rameshbabu <sergeantsagara@protonmail.com>