Handle Pulse main config packets up to 1 MiB
Our implementation has assumed that the entirety of the main Pulse
configuration “packet” will fit in one TLS record; however,
https://gitlab.com/openconnect/openconnect/-/issues/617 demonstrates that it
can in fact exceed 16 KiB if it includes e.g.  a large proxy configuration.

In order to handle this, we need to dynamically allocate the space to hold
this packet, and read it in a loop.

(See https://gitlab.com/openconnect/openconnect/-/commit/2d77040a870851a625de16938fcdda6a5494d7ed
for a previous case where a configuration packet unexpectedly exceeded the
limits of a single TLS record.)

Signed-off-by: Daniel Lenski <dlenski@gmail.com>