Handle '\NNN' escapes in Fortinet search domains
See https://gitlab.com/openconnect/openconnect/-/issues/634 for the details.
Basically, some Fortinet VPNs apparently use '\059' (ASCII character 59 is ';')
as a delimiter in their DNS search domains list.

This *could* be handled in a cross-protocol way, as was done earlier in
https://github.com/dlenski/mariadb-server/commit/03eafb41f45d70938cb19820a21165445ff3af14
for the case of [;, ] separators, but it seems too weird and
Fortinet-specific for me to want to do that immediately.

This also modifies `tests/fake-fortinet-server.py` to delimit search domains
using such escapes, as a test/example for OpenConnect to parse.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>