More tweaks to Pulse and oNCP ESP switching
I could have sworn that at some point this morning, that trick of turning
ESP off and then on again was working, and allowing reconnections. But I
can't reproduce that after correct amounts of coffee. I only ever see it
being turned *off*, and a log message 'Transport mode switched over to SSL
for user with NCIP 10.200.200.1', and staying off.

The server seems to default to ESP enabled if we've configured ESP, even
before the probes occur. The client has to explicitly turn it *off* with
the 'ncmo=0' packet, to fall back to TLS. And then can't turn ESP back
on again, AFAICT. This is kind of stupid, but it doesn't look like we
can do any better.

Signed-off-by: David Woodhouse <dwmw2@infradead.org>