Explicitly specify UTF-8 encoding with htmlReadMemory
Until v2.11.0, the HTML parser in libxml2 defaulted to UTF-8 encoding when
invoked with `htmlReadMemory`.  Apparently, this was actually a bug, or at
least an internal inconsistency, because the HTML parser otherwise defaulted
to ISO-8859-1 (the HTTP protocol's default), so in v2.11.0+ it has been
changed to use ISO-8559-1 as the default.  More details at
https://gitlab.gnome.org/GNOME/libxml2/-/issues/570.

This change in the default interpretation of HTML caused a hard-to-reproduce
regression in OpenConnect's parsing of non-ASCII form fields.  See
https://gitlab.com/openconnect/openconnect/-/issues/642#note_1506117034.

As a short-term fix, OpenConnect should explicitly specify UTF-8 as the
encoding when invoking `htmlReadMemory`.

A correct and complete fix will require us to actually follow the
`Content-Type:` HTTP header when parsing HTML, and probably XML as well.

Signed-off-by: Daniel Lenski <dlenski@gmail.com>