Bump fallback GlobalProtect version number
Some GlobalProtect servers complain about old versions of the client
software connecting to them.

In the case of a connection via the GlobalProtect "portal" interface,
we capture the preferred software version from the portal and parrot it back,
as of https://gitlab.com/openconnect/openconnect/-/commit/c0d2daeaa85f69ed2f89330a53d97ae7eafdffb1?merge_request_iid=333.

However, we should update the GlobalProtect software version used as a fallback
in the case of a direct connection to the "gateway" interface.

Signed-off-by: Dimitri Papadopoulos <3234522+DimitriPapadopoulos@users.noreply.github.com>
Signed-off-by: Daniel Lenski <dlenski@amazon.com>